+++ src/BodyChangeNG/RaceMenuPresetMigration.h
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace bcn::racemenu_preset_migration
+{
+    void MigrateVisiblePresets();
+}